+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file. This projec
 - New `assertDoesntHaveIncluded` assertion to assert that the JSON:API document does not have the top-level `included`
   member.
 - New `assertDoesntHaveMeta` assertion to assert the JSON:API document does not have the top-level `meta` member.
+- New `assertDoesntHaveLinks` assertion to assert the JSON:API document does not have the top-level `links` member.
 
 ### Fixed